requirements.txt
langchain 
pypdf 
pymongo 
langchain-openai 
tiktoken
python-dotenv
langchain_community
langchain_mongodb

